+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [13.3.2] - Unreleased
+## Unreleased
+
+### Fixed
+
+- Fixed pretty printing of empty dataclass https://github.com/Textualize/rich/issues/2819
 
 ### Added
 
@@ -14,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Can now suppress frames using regexes passed via `suppress` parameter to `traceback.Traceback`
+
 
 ## [13.3.1] - 2023-01-28